+++ internal/modules/paths/pops.go
@@ -176,6 +176,7 @@ var pops = map[string]pop{
 	"mad":  {"Madrid, ES", 40.4168, -3.7038},
 	"man":  {"Manchester, GB", 53.4808, -2.2426},
 	"mci":  {"Kansas City, MO, US", 39.0997, -94.5786},
+	"mes":  {"Mesa, AZ, US", 33.4152, -111.8315},
 	"mco":  {"Orlando, FL, US", 28.5383, -81.3792},
 	"mel":  {"Melbourne, AU", -37.8136, 144.9631},
 	"mem":  {"Memphis, TN, US", 35.1495, -90.0490},